pyproject.toml
[project]
dynamic = ["version"]
name = "eips"
description = "Ethereum Improvement Proposal (EIP) ETL library"
authors = [
  { name = "Mike Shultz", email = "mike@mikeshultz.com" },
]
keywords = [
  "ethereum",
  "eip",
  "etl"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Utilities",
]

readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT License"}
dependencies = [
  "click~=8.1",
  "dulwich~=0.21.6",
  "pydantic~=2.3.0",
  "python-dateutil~=2.8.2",
]

[project.optional-dependencies]
dev = [
  "black>=23.9.1",
  "hatch>=1.7.0",
  "isort>=5.12.0",
  "mypy>=1.5.1",
  "pytest>=7.4.2",
  "pytest-cov>=4.1.0",
]

[project.urls]
#Documentation = "..."
"Source code" = "https://github.com/mikeshultz/eips"

[project.scripts]
eips = "eips.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[publish.index.repos.test]
url = "https://test.pypi.org/legacy/"

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true
skip_empty = true

[tool.hatch.envs.default]
# TODO: Windows might work?
platforms = ["linux", "macos"]
features = [
  "dev"
]
python = "3.11"

[tool.hatch.envs.default.scripts]
test = "pytest {args}"
coverage = "test --cov-config=pyproject.toml --cov=eips"
lint = "isort . && black ."

[[tool.hatch.envs.test.matrix]]
python = ["3.8", "3.9", "3.10", "3.11"]
feature = ["dev"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "eips/__init__.py"

[tool.isort]
follow_links = false
profile = "black"
skip_gitignore = true
src_paths = ["eips", "tests"]

[tool.mypy]
plugins = "pydantic.mypy"

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
